加入dtype校验

+++ examples/endless.py
@@ -3,7 +3,7 @@ import numpy as np
 from npyt import NPY8
 
 if __name__ == '__main__':
-    ns = NPY8('demo', 21, 8).load()
+    ns = NPY8('demo', 21, 8, dtype=np.int32).load()
 
     for i in range(9):
         ns.append(np.array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]) + i * 10)
